Fix logic for MEDIA_EJECT intent
When receiving a MEDIA_EJECT broadcast, the VideoModule is calling
stopVideoRecording() no matter what, even though, by default, the video
is saved in primary external storage, which is the emulated sd card.
This is causing the video recording to be stopped when the real sd card
is ejected, even though the video is not saved on that sd card.

In order to fix this issue, this patch checks if the current video path
corresponds to the media ejected.

Change-Id: Ie35f4aad2626cb0a558304018116c187efde0b6e
Signed-off-by: Robert Chiras <robert.chiras@intel.com>