Add scrolling support with a mouse wheel
When a physical mouse is connected, support scroll events
from the mouse when pictures are displayed. In film strip
view (full screen or not), both vertical and horizontal
scroll events have the same bevahior: pictures are
horizontally scrolled. In full screen, pictures are flinged.
When a picture is zoomed the scroll is done either
vertically or horizontally depending on user action.

Change-Id: Idb3b12b0ff86e8a590fa7f1a54111700db84e543
Signed-off-by: Olivier Gay <ogay@logitech.com>