Add scrolling support with a mouse wheel
When a physical mouse is connected, support scroll events
from the mouse when pictures are displayed. In film strip
view (full screen or not), both vertical and horizontal
scroll events have the same bevahior: pictures are
horizontally scrolled. In full screen, pictures are flinged.
When a picture is zoomed the scroll is done either
vertically or horizontally depending on user action.
(Cherry-picked from fa85c84).

Change-Id: I01cad9d114c9f612efe1366a896f5d40f538b71c
Signed-off-by: Olivier Gay <ogay@logitech.com>