Update patch set 1
Abandoned

Let's abandon this change. After discussing with Alan in a separate thread, it appears this porting to Lollipop may have already been done in an internal branch. So this change may come into AOSP through a merge of this internal branch. In any case, Alan if you need this particular change, of course feel free to reopen it.

Patch-set: 1
Status: abandoned