Camera2: correct image size recorded in media provider db
Photos taken in camcorder can't be previewed on PC via MTP mode. The
root cause is that, the actual image file size does not match the
_size column in media provider database. During the saving process,
Jpeg files are compressed, but the size sent to media provider is not
updated accordingly.

Using the saved image file size, the correct image size could be sent
to media provider.

Change-Id: Ic8a92dc854a1fa720dcbd30d83f5772ab9fff170
Signed-off-by: Wang, LisongX <lisongx.wang@intel.com>