Flash icon can not be shown after open and close hdr
The flash icon can't be shown if the flash is not in auto state
and then open and close hdr, for the reason that when we call
the function of syncFlashIndicator, the state of flash icon is
not enable. So we suggest that we can enable the flash icon
firstly and then call syncFlashIndicator to avoid this issue.

Change-Id: I18b684efdf8fe5252dccbfcb2985bfa069b859c2